# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Pokémon Knowledge Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /types/{name}:
    get:
      operationId: getType
      tags: [types]
      summary: Get type details and effectiveness
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Type details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PokemonTypeDetail"

  # ─── Abilities ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    TypeEffectiveness:
      type: object
      properties:
        doubleDamageTo:
          type: array
          items: { type: string }
        halfDamageTo:
          type: array
          items: { type: string }
        noDamageTo:
          type: array
          items: { type: string }
        doubleDamageFrom:
          type: array
          items: { type: string }
        halfDamageFrom:
          type: array
          items: { type: string }
        noDamageFrom:
          type: array
          items: { type: string }
...
    PokemonTypeDetail:
      type: object
      required: [id, name, color, effectiveness]
      properties:
        id: { type: integer }
        name: { type: string }
        color: { type: string }
        effectiveness:
          $ref: "#/components/schemas/TypeEffectiveness"
        pokemonCount: { type: integer }

    # ── Abilities ───────────────────────────────────────────────────────────
...