# lib/api-zod/src/generated/types/pokemonTypeDetail.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Pokémon Knowledge Management Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { TypeEffectiveness } from './typeEffectiveness';

export interface PokemonTypeDetail {
  id: number;
  name: string;
  color: string;
  effectiveness: TypeEffectiveness;
  pokemonCount?: number;
}